requirements.txt
nibabel==5.3.2
numpy==1.23.5
numpy==1.24.4
PyYAML==5.4.1
torch==2.3.1+cu118
tqdm==4.66.4
